==> 6_model_building.Rmd <==
---
title: "6.2_model_building"
author: "Borka Domsodi"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(message=FALSE)

# Set working directory
setwd("~/my_computer/my_folder") # Replace with local folder path 

# Load packages
library(readxl)
library(tidyverse)
library(modelsummary)
library(stargazer)
library(pscl)
```

```{r}
# Load Data
load("5.4_variable_inspection_output.RData")
```

### MODEL 1: BASELINE MODEL ###

```{r}
# Model 1
model1 <- glm(label_dummy ~ social_support_index, data = analysis_complete, family = "binomial")
summary(model1)
```

```{r}
# Save the model summary output in a presentable format
stargazer(model1,
          type = "html",
                     align = FALSE,
                     single.row = TRUE,
                     df = FALSE,
                     star.cutoffs = c(0.05, 0.01, 0.001),
                     intercept.bottom = FALSE,
                     dep.var.labels = "Recovery",
                     covariate.labels = c("Intercept: non-recovered",
                                          "Social Support"),
                     title = "Baseline Model",
                     keep.stat = c("n","f"),
                     no.space = TRUE,
                     out = "6.1.1_model1.htm")
```

```{r}
# Calculating pseudo R squared
pseudo <- pR2(model1)
print(paste("The baseline model explains around", round(pseudo[4], digits=4)*100, "% of the total variance of the Recovery variable."))
rm(pseudo)
```

```{r}
# Plotting the probability curve

# Create dataframe with probabilities and actual survival
predicted.data <- data.frame(probability.of.label_dummy=model1$fitted.values, recovery=analysis_complete$label)
predicted.data <- predicted.data[order(predicted.data$probability.of.label_dummy, decreasing=FALSE),]
predicted.data$rank <- 1:nrow(predicted.data)
 
# Plot probability curve
ggplot(data=predicted.data, aes(x=rank, y=probability.of.label_dummy)) +
  geom_point(aes(color=recovery)) +
  theme_minimal() + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        text = element_text(size = 12, family = "serif"),
        legend.title = element_blank()) + 
  labs(x = "
       Index",
       y = "Predicted probability of Recovery
                                            ",
       title = "Probability Plot - Baseline Model") + 
  coord_cartesian(ylim = c(0, 1))
rm(predicted.data)
ggsave("6.2.1_probability1.jpg")
```

```{r}
# Assumptions

# The deviance residuals of this model are acceptable as they are close to being centered on 0 and are roughly symmetrical which suggests that they are normally distributed

# Residual Plot
plot(model1, 5)
plot(model1, 4)
# No influential outliers are detected on the Residuals vs Leverage plot
# Looking at the Cook's distance plot, no highly influential outliers are detected

# The degrees of freedom is high enough

# Clean Environment
rm(model1)
```

### MODEL 2: ELABORATION MODEL ###

```{r}
# Standardize the indices so the coefficients are commensurable
analysis_complete <- analysis_complete %>% 
  mutate(st_social_integration_index = scale(social_integration_index),
         st_social_network_structure_index = scale(social_network_structure_index),
         st_relational_content = scale(sentiment_score),
         st_outbound_sentiment_score = scale(outbound_sentiment_score))

# Model 2
model2 <- glm(label_dummy ~ st_social_integration_index + st_social_network_structure_index + st_relational_content, data = analysis_complete, family = "binomial")
summary(model2)
```

```{r}
# Save the model summary output in a presentable format
stargazer(model2,
          type = "html",
                     align = FALSE,
                     single.row = TRUE,
                     df = FALSE,
                     star.cutoffs = c(0.05, 0.01, 0.001),
                     intercept.bottom = FALSE,
                     dep.var.labels = "Recovery",
                     covariate.labels = c("Intercept: non-recovered",
                                          "Social Integration",
                                          "Social Network Structure",
                                          "Relational Content"),
                     title = "Elaboration Model",
                     keep.stat = c("n","f"),
                     no.space = TRUE,
                     out = "6.1.2_model2.htm")
```

```{r}
# Calculating pseudo R squared
pseudo <- pR2(model2)
print(paste("The elaboration model explains around", round(pseudo[4], digits=4)*100, "% of the total variance of the Recovery variable."))
rm(pseudo)
```

```{r}
# Plotting the probability curve

# Create dataframe with probabilities and actual survival
predicted.data <- data.frame(probability.of.label_dummy=model2$fitted.values, recovery=analysis_complete$label)
predicted.data <- predicted.data[order(predicted.data$probability.of.label_dummy, decreasing=FALSE),]
predicted.data$rank <- 1:nrow(predicted.data)
 
# Plot probability curve
ggplot(data=predicted.data, aes(x=rank, y=probability.of.label_dummy)) +
  geom_point(aes(color=recovery)) +
  theme_minimal() + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        text = element_text(size = 12, family = "serif"),
        legend.title = element_blank()) + 
  labs(x = "
       Index",
       y = "Predicted probability of Recovery
                                            ",
       title = "Probability Plot - Elaboration Model") + 
  coord_cartesian(ylim = c(0, 1))
rm(predicted.data)
ggsave("6.2.2_probability2.jpg")
```

```{r}
# Assumptions

# The deviance residuals of this model are acceptable as they are close to being centered on 0 and are roughly symmetrical which suggests that they are normally distributed

# Residual Plot
plot(model2, 5)
plot(model2, 4)
# No influential outliers are detected on the Residuals vs Leverage plot
# Looking at the Cook's distance plot, no highly influential outliers are detected

# The degrees of freedom is high enough

# Clean Environment
rm(model2)
```

### MODEL 3: EXCLUSIONARY MODEL ###

```{r}
# Model 3
model3 <- glm(label_dummy ~ st_social_integration_index + st_social_network_structure_index + st_relational_content + st_outbound_sentiment_score, data = analysis_complete, family = "binomial")
summary(model3)

# Calculating pseudo R squared
pseudo <- pR2(model3)
print(paste("Pseudo R-squared:", round(pseudo[4], digits=4)*100, "%"))
rm(pseudo)
```

```{r}
# Save the model summary output in a presentable format
stargazer(model3,
            type = "html",
            align = FALSE,
            single.row = TRUE,
            df = FALSE,
            star.cutoffs = c(0.05, 0.01, 0.001),
            intercept.bottom = FALSE,
            dep.var.labels = "Recovery",
            covariate.labels = c("Intercept: non-recovered",
                                 "Social Integration",
                                 "Social Network Structure",
                                 "Relational Content",
                                 "Confounder: Outbound Relational Content"),
            title = "Exclusionary Model",
            keep.stat = c("n","f"),
            no.space = TRUE,
            out = "6.1.3_model3.htm")
```

```{r}
# Plotting the probability curve

# Create dataframe with probabilities and actual survival
predicted.data <- data.frame(probability.of.label_dummy=model3$fitted.values, recovery=analysis_complete$label)
predicted.data <- predicted.data[order(predicted.data$probability.of.label_dummy, decreasing=FALSE),]
predicted.data$rank <- 1:nrow(predicted.data)
 
# Plot probability curve
ggplot(data=predicted.data, aes(x=rank, y=probability.of.label_dummy)) +
  geom_point(aes(color=recovery)) +
  theme_minimal() + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        text = element_text(size = 12, family = "serif"),
        legend.title = element_blank()) + 
  labs(x = "
       Index",
       y = "Predicted probability of Recovery
                                            ",
       title = "Probability Plot - Exclusionary Model") + 
  coord_cartesian(ylim = c(0, 1))
rm(predicted.data)
ggsave("6.2.3_probability3.jpg")
```

```{r}
# Assumptions

# The deviance residuals of this model are acceptable as they are close to being centered on 0 and are roughly symmetrical which suggests that they are normally distributed

# Residual Plot
plot(model3, 5)
plot(model3, 4)
# No influential outliers are detected on the Residuals vs Leverage plot
# Looking at the Cook's distance plot, no highly influential outliers are detected

# The degrees of freedom is high enough

# Clean Environment
rm(model3)
```

```{r}
# Clean Global Environment
rm(list = ls())
```
